+++ .github/instructions/copilot-instructions.md
@@ -1,0 +1,6 @@
+---
+applyTo: '**'
+---
+このサイトは、tailwindv4です。v3ではないことに注意。
+lintの、改行やスペース, importのフォーマットのルールでのエラーは完全に無視して良い。自動修正を拡張機能がしてくれるから。
+静的サイトです.
