+++ package.json
@@ -19,17 +19,27 @@
     "@codemirror/theme-one-dark": "^6.1.2",
     "@codemirror/view": "^6.34.1",
     "@tabler/icons-react": "^3.34.1",
+    "@types/katex": "^0.16.7",
+    "@types/react-syntax-highlighter": "^15.5.13",
     "@uiw/react-codemirror": "^4.23.5",
     "@xterm/addon-fit": "^0.10.0",
     "@xterm/xterm": "^5.5.0",
     "firebase": "^12.2.1",
     "framer-motion": "^12.23.12",
     "gsap": "^3.13.0",
+    "katex": "^0.16.22",
     "lucide-react": "^0.542.0",
+    "mermaid": "^11.11.0",
     "pyodide": "^0.26.4",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-router-dom": "^7.8.2"
+    "react-markdown": "^10.1.0",
+    "react-router-dom": "^7.8.2",
+    "react-syntax-highlighter": "^15.6.6",
+    "rehype-highlight": "^7.0.2",
+    "rehype-katex": "^7.0.1",
+    "remark-gfm": "^4.0.1",
+    "remark-math": "^6.0.0"
   },
   "devDependencies": {
     "@eslint/compat": "^1.2.9",
