+++ src/index.css
@@ -1,6 +1,13 @@
 @import "tailwindcss";
 /* @import "tailwindcss/typography"; */
 
+/* KaTeX CSS */
+@import "katex/dist/katex.min.css";
+
+/* Highlight.js CSS */
+@import "highlight.js/styles/github.css";
+@import "highlight.js/styles/github-dark.css" (prefers-color-scheme: dark);
+
 @layer base {
   html,
   body {
